# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Bruma API
  version: 1.0.0
  description: |
    MVP HTTP API (tenants, domains, templates, dispatch, webhooks, logs). Authenticate with `X-API-Key`
    (admin key or per-tenant key). Webhook endpoint uses HMAC signature headers instead of API keys.

    **Swagger UI:** `GET /docs` when `OPENAPI_DOCS_ENABLED` is true or outside production (see deployment config).
servers:
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /v1/tenants/{tenantId}/email-dispatches/{dispatchId}:
    get:
      tags: [Dispatch]
      summary: Get dispatch row by id
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - name: dispatchId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchDetail" }
  /v1/dispatch:
    post:
      tags: [Dispatch]
      summary: Enqueue email dispatch (requires Redis / worker)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DispatchRequest" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [message_id, status]
                    properties:
                      message_id: { type: string, format: uuid }
                      status: { type: string }
                  - type: object
                    required: [message_id, status, idempotent]
                    properties:
                      message_id: { type: string, format: uuid }
                      status: { type: string }
                      idempotent: { type: boolean, const: true }
        "503": { description: REDIS_URL not configured }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    TenantIdPath:
      name: tenantId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Dispatch:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        status: { type: string }
        event_key: { type: [string, "null"] }
        recipient_email: { type: string }
        template_version_id: { type: [string, "null"], format: uuid }
        sender_id: { type: [string, "null"], format: uuid }
        subject: { type: [string, "null"] }
        provider_message_id: { type: [string, "null"] }
        correlation_id: { type: [string, "null"] }
        last_error: { type: [string, "null"] }
        idempotency_key: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DispatchDetail:
      allOf:
        - { $ref: "#/components/schemas/Dispatch" }
        - type: object
          properties:
            events:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  event_type: { type: string }
                  occurred_at: { type: string, format: date-time }
    DispatchRequest:
      type: object
      required: [tenant_id, event, recipient]
      properties:
        tenant_id: { type: string, format: uuid }
        event: { type: string }
        recipient:
          type: object
          required: [email]
          properties:
            email: { type: string, format: email }
        variables:
          type: object
          additionalProperties: true
        idempotency_key: { type: string }
...
security:
  - ApiKeyAuth: []